+++ constants/index.js
@@ -1,0 +1,5 @@
+const { YEARTEXT, JW_API, JWPUB_API, REPO_URL, FFMPEG_VERSION } = require('./endpoints')
+
+module.exports = {
+  YEARTEXT, JW_API, JWPUB_API, REPO_URL, FFMPEG_VERSION
+}
